Re-blacken everything
- Reformats everything to be consistent with `black` tool
- Blacken was run in the past and after that, there was some period
  without having travis blacken-check the changes. This resulted in
inconsistencies were a contributor's code diff is black-checked by
travis whereas at the same time there are many parts of the code base
that are not "blackened". So if a contributor runs black locally they
end up with a huge amount of reformatting changes unrelated to their
changes. This commits resets everything to the formatting convention of
`black` and with the travis check already set there won't be such issue
in the future